+++ libevmasm/GasMeter.cpp
@@ -129,6 +129,7 @@ GasMeter::GasConsumption GasMeter::estimateMax(AssemblyItem const& _item, bool _
 		case Instruction::CALL:
 		case Instruction::CALLCODE:
 		case Instruction::DELEGATECALL:
+		case Instruction::STATICCALL:
 		{
 			if (_includeExternalCosts)
 				// We assume that we do not know the target contract and thus, the consumption is infinite.
@@ -142,8 +143,10 @@ GasMeter::GasConsumption GasMeter::estimateMax(AssemblyItem const& _item, bool _
 					gas = GasConsumption::infinite();
 				if (_item.instruction() == Instruction::CALL)
 					gas += GasCosts::callNewAccountGas; // We very rarely know whether the address exists.
-				int valueSize = _item.instruction() == Instruction::DELEGATECALL ? 0 : 1;
-				if (!classes.knownZero(m_state->relativeStackElement(-1 - valueSize)))
+				int valueSize = 1;
+				if (_item.instruction() == Instruction::DELEGATECALL || _item.instruction() == Instruction::STATICCALL)
+					valueSize = 0;
+				else if (!classes.knownZero(m_state->relativeStackElement(-1 - valueSize)))
 					gas += GasCosts::callValueTransferGas;
 				gas += memoryGas(-2 - valueSize, -3 - valueSize);
 				gas += memoryGas(-4 - valueSize, -5 - valueSize);